refactor: useWebviewCamera hook
